requirements.txt
fastapi
uvicorn
python-multipart
pandas
joblib
openai
git+https://github.com/openai/whisper.git
python-dotenv

